Remove use of deprecated $.ucFirst()
This is the last instance of $.ucFirst() in Wikimedia Git.

Usually inlined as `x[ 0 ].toUpperCase() + x.slice( 1 )`, but in this case
the call can simply be removed as it is not used for display but simply
passed to resolveRedirects() which already normalises it through mw.Title.
No need for the input to be semi-normalised here.

Change-Id: I616d03a256a5a635472c9531f2a5ab84e19b9b68