Fix RegExp escaping of first character
Use charAt( 0 ) from original string instead of the escaped string.

If the first character of the original string contains a special RegExp
character then charAt( 0 ) of the escaped string contains '\' instead
of the original first character of the category.

The RegExp escaping has to be done after the split of the string.

Change-Id: I8b715963e2a242fffb902a002b03ddd26859b622