Remove needless use of mw.util.$content
When `dialog()` is called to (auto-)open the dialog, the $dialog
content will be attached to the right place in the jQuery UI dialog.

There is no need to first append it somewhere else random, as was
done in mw.util.$content, one statement before it is put elsewhere.

Test Plan:
* Put `wfLoadExtension('InlineCategorizer');` in LocalSettings.php.
* As an example to trigger showError, modify the handleCategoryAdd()
  function and make its condition `if ( true )` so that pressing "Add"
  will always end in an error.
* View the Main Page, click "Add category", type "Bla", click "Add".
* Observe the error still being displayed correctly and in the
  same way as without this change.

Bug: T65466
Change-Id: Iac2011db045c9bf9caa38e19c1d3dd3ea3c35416